pyproject.toml
[tool.poetry]
name = "desdeo-webapi"
version = "0.1.0"
description = ""
authors = ["Giovanni Misitano <giovanni.a.misitano@jyu.fi>"]

[tool.poetry.dependencies]
python = ">=3.9, < 3.11"
Flask = "^2.2"
requests = "^2.25.1"
Flask-JWT-Extended = "^4.0.2"
Flask-SQLAlchemy = ">=3.0.2"
passlib = "^1.7.4"
dill = {git = "https://github.com/uqfoundation/dill",  branch = "master" }
sympy = "^1.7.1"
Flask-Cors = "^3.0.10"
desdeo-problem = "^1.3.3"
flask-restx = "^1.0.0"
pandas = "^1.3.1"
simplejson = "^3.17.3"
openpyxl = "^3.0.9"
Werkzeug = ">=2.2"
desdeo-mcdm = "^1.3.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
rope = "^0.18.0"
Flask-Testing = "^0.8.1"
Sphinx = "<=6.2.1"
sphinxcontrib-httpdomain = "^1.7.0"
rstcheck = "^3.3.1"
sphinx-rtd-theme = "^0.5.1"
pytest = "^6.2.3"
matplotlib = "^3.4.2"

[tool.poetry.group.dev.dependencies]
pytest-rerunfailures = "^11.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
